refactor: split into sub components
